+++ full-stack-contactMgmtSys/src/main/resources/application.properties
@@ -12,3 +12,4 @@ spring.datasource.password=${mysqlpassword}
 spring.jpa.show-sql=true
 #spring.jpa.properties.hibernate.dialect=org.hibernate.dialect.MySQLDialect
 spring.jpa.hibernate.ddl-auto=update
+
